hop-node: add confirmRoot command to bypass canonical messenger
