+++ packages/hop-node/src/cli/cli.ts
@@ -7,9 +7,9 @@ import './bondTransferRoot'
 import './bondWithdrawal'
 import './challenger'
 import './commitTransfers'
+import './confirmRoot'
 import './contractState'
 import './dbDump'
-import './exitCommitTx'
 import './healthCheck'
 import './hopNode'
 import './incompleteSettlements'
